Make a recipes directory and add one recipe file
Make a new directory called recipes that will contain all the recipes in the website.
This directory was made to make the file structure organized.

In this new directory, create the first recipe file under the name doroWat.html along
with the usual html boilerplate and a heading that indicates the dish's name.